[coqlib] Rebindable Coqlib namespace.
We refactor the `Coqlib` API to locate objects over a namespace
`module.object.property`.

In the core, we deprecate all the old functions; also we discard
"lazy" binding in favor of truly dynamic one.

For plugins, we still allow `coq_reference` until we provide the
rewindable objects.

There is still much more work to do to get proper rebindable objects,
but this should be a good step.