Support int63 in numeral notations
This commit allows users to use `Int63.int` as a source and/or target in
numeral notations printing and parsing functions.  This is convenient
for any user number type based on int63, so that a round-trip with `Z`
or `Decimal.int` is not required.

Additionally, this makes int63_syntax.ml significantly shorter (by
moving most of the interesting work on the Int constructor in the kernel
constr AST into numeral interpretations, where it belongs).  Note that
this requires adding an identity function specialized to `Int63.int`
into the `Int63.v` file, as numeral notations require the printing and
parsing functions to be single identifiers.

Note that this paves the way for moving the int63 syntax declaration
into user-space.  There are only two remaining issues:
1. Numeral Notation will have to be extended to accept primitive types
   and not just inductive types as the binding types for numeral
   notations.
2. The printing infrastructure around the Int constructor will have to
   be updated so that it goes through numeral notations; as it is
   currently, the printing of primitive integers is hard-coded into the
   printing mechanism.

Regardless of whether or not we eventually move int63 syntax into
user-space, I think this change is worth it to support numeral parsing
and printing of int63-based types more conveniently.

Note that for the quick build of int63 syntax, we must wait until ML
module declaration time to locate the int63 function.